pyproject.toml
[project]
name = "nestory"
version = "0.1.0"
description = "은퇴자 전원주택 커뮤니티 웹앱"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "pydantic-settings>=2.5",
  "jinja2>=3.1",
  "python-multipart>=0.0.12",
  "email-validator>=2.2",
  "itsdangerous>=2.2",
  "argon2-cffi>=23.1",
  "httpx>=0.27",
  "structlog>=24.4",
  "sentry-sdk[fastapi]>=2.17",
  "pillow>=11.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "factory-boy>=3.3",
  "ruff>=0.7",
  "mypy>=1.13",
  "types-itsdangerous",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
# B008: FastAPI idiom — Depends() in default args is intentional.
# UP042: str+Enum is deliberate for PG enum value-based storage (see app/models/user.py).
ignore = ["B008", "UP042"]

[tool.ruff.lint.per-file-ignores]
"app/db/migrations/versions/*.py" = ["E501", "W291", "I001", "UP035"]
"app/db/migrations/env.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"
